PPApiHooks + PlusPrivacy
+ Temporarily disabled the swizzling of CNContactStore and
UIImagePickerController in PPApiHooks.framework;. It seems that the
presence of these symbols trigger a warning on the binary validation,
requiring the presence of "NSContactsUsageDescription"
"NSPhotoLibraryUsageDescription" keys in the Info.plist file.

This requires some changes in how the libraries are to be compiled and
sent to the app developers. Must investigate