ApiHooks separation
"Hooking" code for classes / items which require special permissions
via a message key in the Info.plist (ContactStore, LocalAuthentication,
etc..) must be kept in separate frameworks. Thus there will be a "core"
hooking library, containing the event dispatcher + other hooked classes
which do not require special permissions (CMMotionManager, UIDevice,
etc..). But this leaves the PPEventDispatcher class public and open for
everyone (it must be public in order for the classes in the other
libraries to register a callback). Must find a way to protect it and
allow/ validate access to it for certain callers only.